+++ src/helpers.ts
@@ -16,6 +16,9 @@
  */
 
 import { _Connection } from "vscode-languageserver/node";
+import { PSSLangObjects } from "./definitions/dataStructures";
+import { updateASTNew, updateASTNewMeta } from "./parser/helpers";
+import fs from 'fs-extra';
 
 export function notify(conn: _Connection, message: string, err: boolean = false) {
   if (err) {
@@ -23,4 +26,19 @@ export function notify(conn: _Connection, message: string, err: boolean = false)
   } else {
     conn.window.showInformationMessage(message);
   }
+}
+
+export function buildASTForFiles(files: string[]): PSSLangObjects[]{
+  let pssAST: PSSLangObjects[] = [];
+  for (const file of files) {
+    const content: string = fs.readFileSync(file, 'utf8');
+    const fileURI: string = encodeURI("file://" + file);
+    // updateAST(fileURI, content).then(vars => {
+    //   globalAST = updateASTMeta(globalAST, vars);
+    // });
+    updateASTNew(fileURI, content).then(vars => {
+      pssAST = updateASTNewMeta(pssAST, vars);
+    });
+  }
+  return pssAST;
 }